+++ R/tableby.R
@@ -353,11 +353,11 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, strata,
             for(bylev in by.levels) {
               idx <- bycol == bylev
               bystatlist[[bylev]] <- do.call(statfun, list(currcol[idx], levels=xlevels, na.rm=TRUE,
-                                                           weights=weightscol[idx], conf.level = control$conf.level))
+                                                           weights=weightscol[idx], conf.level=control$conf.level, times=control$times))
             }
             ## add Total
             bystatlist$Total <- do.call(statfun, list(currcol, levels=xlevels, na.rm=TRUE,
-                                                      weights=weightscol, conf.level = control$conf.level))
+                                                      weights=weightscol, conf.level=control$conf.level, times=control$times))
           }
           statList[[statfun]] <- bystatlist
         }